+++ src/TinyGsmClientSIM808.h
@@ -63,7 +63,7 @@ class TinyGsmSim808 : public TinyGsmSim800, public TinyGsmGPS<TinyGsmSim808> {
       float ilat         = 0;
       float ilon         = 0;
       float ispeed       = 0;
-      int   ialt         = 0;
+      float ialt         = 0;
       int   ivsat        = 0;
       int   iusat        = 0;
       float iaccuracy    = 0;
@@ -105,7 +105,7 @@ class TinyGsmSim808 : public TinyGsmSim800, public TinyGsmGPS<TinyGsmSim808> {
       if (lat != NULL) *lat = ilat;
       if (lon != NULL) *lon = ilon;
       if (speed != NULL) *speed = ispeed;
-      if (alt != NULL) *alt = ialt;
+      if (alt != NULL) *alt = static_cast<int>(ialt);
       if (vsat != NULL) *vsat = ivsat;
       if (usat != NULL) *usat = iusat;
       if (accuracy != NULL) *accuracy = iaccuracy;
